print out types of variable

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ long int longIntType;
 long long int longLongIntType;
 float floatType;
 
-printf("Size of a char: %zu bytes\n", sizeof(charType));
-printf("Size of an int: %zu bytes\n", sizeof(intType));
-printf("Size of a long int: %zu bytes\n", sizeof(longIntType));
-printf("Size of a long long int: %zu bytes\n", sizeof(longLongIntType));
-printf("Size of a float: %zu bytes\n", sizeof(floatType));
+printf("Size of a char: %lu bytes\n", (unsigned long)sizeof(charType));
+printf("Size of an int: %lu bytes\n", (unsigned long)sizeof(intType));
+printf("Size of a long int: %lu bytes\n", (unsigned long)sizeof(longIntType));
+printf("Size of a long long int: %lu bytes\n", (unsigned long)sizeof(longLongIntType));
+printf("Size of a float: %lu bytes\n", (unsigned long)sizeof(floatType));
 
 return (0);
 }
